chore. 변수 명 변경
