+++ src/components/Layout.jsx
@@ -3,7 +3,7 @@ import { css } from '@emotion/react';
 import { Outlet } from 'react-router-dom';
 import Navbar from './Navbar';
 import Footer from './Footer';
-import HeaderComponent from './HeaderComponent';
+import Header from './Header';
 
 const Layout = () => {
   return (
@@ -19,7 +19,7 @@ const Layout = () => {
           box-sizing: border-box;
         `}
       >
-        <HeaderComponent />
+        <Header />
         <Outlet />
       </div>
       <Footer />
